Music pause/resume for a phone call/homescreen button exit
Music should pause and resume appropriately following a phone call or
exiting the app using the homescreen button. This does not handle a user
pushing the power button while in game (turned out to be a big pain
because our MainActivity is in landscape and the lockscreen is in
portrait). Currently MainActivity is just recreated when a user returns
from pushing the power button.